+++ apps/hub/src/squiggle/components/EditSquiggleSnippetModel.tsx
